+++ format/FormatNumber.js
@@ -35,6 +35,9 @@ const UP_TICK = '▴',
  *      align vertically with negative ledgers in columns.
  * @param {boolean} [opts.withPlusSign] - true to prepend positive numbers with a '+'.
  * @param {boolean} [opts.withSignGlyph] - true to prepend an up / down arrow.
+ * @param {boolean} [opts.withCommas] - true to include comma delimiters.
+ * @param {boolean} [opts.omitFourDigitComma] - set this to true to omit comma if value has exactly
+ *      4 digits (i.e. 1500 instead of 1,500).
  * @param {string?} [opts.prefix] - prefix to prepend to value (between the number and its sign).
  * @param {string?} [opts.label] - label to append to value.
  * @param {string} [opts.labelCls] - CSS class of label <span>,
@@ -60,6 +63,8 @@ export function fmtNumber(v, {
     forceLedgerAlign = true,
     withPlusSign = false,
     withSignGlyph = false,
+    withCommas = true,
+    omitFourDigitComma = false,
     prefix = null,
     label = null,
     labelCls = 'xh-units-label',
@@ -71,7 +76,7 @@ export function fmtNumber(v, {
 
     if (isInvalidInput(v)) return nullDisplay;
 
-    formatConfig = formatConfig || buildFormatConfig(v, precision, zeroPad);
+    formatConfig = formatConfig || buildFormatConfig(v, precision, zeroPad, withCommas, omitFourDigitComma);
     const str = numbro(v).format(formatConfig).replace('-', '');
     let sign = null;
 
@@ -312,10 +317,10 @@ function valueColor(v, colorSpec) {
     return colorSpec.neutral;
 }
 
-function buildFormatConfig(v, precision, zeroPad) {
+function buildFormatConfig(v, precision, zeroPad, withCommas, omitFourDigitComma) {
     const num = Math.abs(v);
 
-    const config = {thousandSeparated: num >= 1000};
+    const config = {};
     let mantissa = undefined;
 
     if (precision % 1 === 0) {
@@ -334,6 +339,9 @@ function buildFormatConfig(v, precision, zeroPad) {
             mantissa = 0;
         }
     }
+
+    config.thousandSeparated = withCommas &&
+        !(omitFourDigitComma && num < 10000 && (mantissa == 0 || (!zeroPad && Number.isInteger(num))));
     config.mantissa = mantissa;
     config.trimMantissa = !zeroPad && mantissa != 0;
     return config;